requirements.txt
django
pillow
django-crispy-forms
django-ckeditor
django-import-export
crispy-bootstrap5
django-utils-six
psycopg == 3.1.8
